Add files via upload
The program has a user interface where the user can enter a keyword to search for on www.seznamzpravy.cz. The program uses web scraping techniques to crawl the website and extract the relevant articles based on the keyword entered by the user. Once the program has scraped the articles, it gives the user the option to select which articles they want to repost to their Facebook account or page. The program then uses Facebook's API to post the selected articles to the user's account or page.